Lab3_4/requirements.txt
confluent-kafka==2.6.0
Flask==3.0.3
PyYAML==6.0.2
